feat: add svelte query
