+++ package.json
@@ -20,6 +20,7 @@
 	},
 	"dependencies": {
 		"@floating-ui/dom": "1.6.11",
+		"@tanstack/svelte-query": "^5.59.0",
 		"bcryptjs": "^2.4.3",
 		"cuid": "^3.0.0",
 		"highlight.js": "11.10.0",
