+++ src/routes/profile-generator/ProfileCard.svelte
@@ -1,5 +1,6 @@
 <script lang="ts">
 	import { createEventDispatcher } from 'svelte';
+	import { QueryClient, QueryClientProvider, createQuery } from '@tanstack/svelte-query';
 
 	export let cuid: string;
 	export let title: string;
@@ -11,17 +12,49 @@
 	let statusColor: string;
 	const dispatch = createEventDispatcher();
 
-	switch (status) {
-		case 'posted':
-			statusColor = 'variant-filled-success';
-			break;
-		case 'received':
-		case 'validated':
-			statusColor = 'variant-filled-warning';
-			break;
-		default:
-			statusColor = 'variant-filled-error';
-			break;
+	const queryClient = new QueryClient();
+
+	// Function to fetch status
+	async function fetchStatus(node_id: string): Promise<string> {
+		if (node_id === '') return 'unknown';
+
+		try {
+			const response = await fetch(`/profile-generator/index/${node_id}`);
+			if (response.ok) {
+				const data = await response.json();
+				return data.status ?? 'unknown';
+			} else {
+				console.error('Failed to fetch status:', response.statusText);
+				return 'unknown';
+			}
+		} catch (error) {
+			console.error('Error fetching status:', error);
+			return 'unknown';
+		}
+	}
+
+	// Use svelte-query to fetch status
+	const statusQuery = createQuery({
+		queryKey: ['status', node_id],
+		queryFn: () => fetchStatus(node_id),
+		refetchInterval: 5000
+	});
+
+	$: {
+		status = $statusQuery.data ?? status;
+
+		switch (status) {
+			case 'posted':
+				statusColor = 'variant-filled-success';
+				break;
+			case 'received':
+			case 'validated':
+				statusColor = 'variant-filled-warning';
+				break;
+			default:
+				statusColor = 'variant-filled-error';
+				break;
+		}
 	}
 
 	function handleDelete() {
@@ -65,23 +98,25 @@
 	}
 </script>
 
-<div class="card variant-ghost-primary border-2 mx-2 my-4 p-4">
-	<div class="font-medium">{title}</div>
-	<div class="m-4">
-		<span class="badge {statusColor} font-bold text-sm mx-4">{status}</span>
-		<span class="badge variant-ghost-primary font-bold text-sm mx-4 mt-2">{last_updated}</span>
+<QueryClientProvider client={queryClient}>
+	<div class="card variant-ghost-primary border-2 mx-2 my-4 p-4">
+		<div class="font-medium">{title}</div>
+		<div class="m-4">
+			<span class="badge {statusColor} font-bold text-sm mx-4">{status}</span>
+			<span class="badge variant-ghost-primary font-bold text-sm mx-4 mt-2">{last_updated}</span>
+		</div>
+		<div class="flex justify-center">
+			<ul class="list text-xs">
+				{#each schemas as schema}
+					<li>{schema}</li>
+				{/each}
+			</ul>
+		</div>
+		<div class="flex justify-around mt-4 md:mt-8">
+			<button class="btn font-semibold md:btn-lg variant-filled-primary">Modify</button>
+			<button on:click={handleDelete} class="btn font-semibold md:btn-lg variant-filled-secondary"
+				>Delete</button
+			>
+		</div>
 	</div>
-	<div class="flex justify-center">
-		<ul class="list text-xs">
-			{#each schemas as schema}
-				<li>{schema}</li>
-			{/each}
-		</ul>
-	</div>
-	<div class="flex justify-around mt-4 md:mt-8">
-		<button class="btn font-semibold md:btn-lg variant-filled-primary">Modify</button>
-		<button on:click={handleDelete} class="btn font-semibold md:btn-lg variant-filled-secondary"
-			>Delete</button
-		>
-	</div>
-</div>
+</QueryClientProvider>
